refactor(front): move conversation DTO types out of lib/api into types/api
Third batch of the lib/api DTO migration: the assistant/conversation domain
(internal endpoints only). Transport request/response DTO types and zod schemas
move to types/api/assistant/conversation/* (mirroring the lib/api module path),
leaving business-logic functions and domain types in place.

- 3 pure-contract files moved wholesale: search, spaces, wakeups.
- 3 files trimmed (DTOs extracted, logic/domain stay): forks, mentions,
  semantic_search. semantic_search.ts keeps the business constant
  SEMANTIC_SEARCH_SCORE_CUTOFF; mentions.ts drops its now-unused zod import.
- 16 consumers updated across front and front-api; one mixed-import route split
  (lib for the business const, types for the schema/DTO).

conversation/types.ts is intentionally excluded: it has a public v1 consumer
(PatchConversationResponseBody) and belongs in the SDK, handled in a later pass.
