+++ front-api/routes/w/[wId]/assistant/conversations/[cId]/wakeups/index.ts
@@ -1,6 +1,6 @@
-import type { GetConversationWakeUpsResponseBody } from "@app/lib/api/assistant/conversation/wakeups";
 import { ConversationResource } from "@app/lib/resources/conversation_resource";
 import { WakeUpResource } from "@app/lib/resources/wakeup_resource";
+import type { GetConversationWakeUpsResponseBody } from "@app/types/api/assistant/conversation/wakeups";
 import { apiErrorForConversation } from "@front-api/lib/api/assistant/conversation/helper";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
