+++ front-api/routes/w/[wId]/assistant/conversations/spaces/index.ts
@@ -1,7 +1,7 @@
-import type { GetBySpacesSummaryResponseBody } from "@app/lib/api/assistant/conversation/spaces";
 import { listNonArchivedMemberSpacesWithMetadata } from "@app/lib/api/projects/list";
 import { ConversationResource } from "@app/lib/resources/conversation_resource";
 import { UserProjectPreferencesResource } from "@app/lib/resources/user_project_preferences_resource";
+import type { GetBySpacesSummaryResponseBody } from "@app/types/api/assistant/conversation/spaces";
 import type { ConversationWithoutContentType } from "@app/types/assistant/conversation";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
