+++ front/components/assistant/conversation/sidebar/PodList.tsx
@@ -3,11 +3,11 @@ import { SidebarContext } from "@app/components/sparkle/SidebarContext";
 import { useConversation } from "@app/hooks/conversations";
 import { usePodConversations } from "@app/hooks/conversations/usePodConversations";
 import { useActiveConversationId } from "@app/hooks/useActiveConversationId";
-import type { GetBySpacesSummaryResponseBody } from "@app/lib/api/assistant/conversation/spaces";
 import { useAppRouter } from "@app/lib/platform";
 import { getSpaceIcon } from "@app/lib/spaces";
 import { removeDiacritics, subFilter } from "@app/lib/utils";
 import { getPodRoute } from "@app/lib/utils/router";
+import type { GetBySpacesSummaryResponseBody } from "@app/types/api/assistant/conversation/spaces";
 import type { ConversationWithoutContentType } from "@app/types/assistant/conversation";
 import type { PodType } from "@app/types/space";
 import type { WorkspaceType } from "@app/types/user";
