+++ front/hooks/conversations/usePodConversations.ts
@@ -1,14 +1,14 @@
-import type {
-  GetBySpacesSummaryResponseBody,
-  GetSpaceConversationsResponseBody,
-  GetSpaceUnreadConversationsResponseBody,
-} from "@app/lib/api/assistant/conversation/spaces";
 import {
   emptyArray,
   useFetcher,
   useSWRInfiniteWithDefaults,
   useSWRWithDefaults,
 } from "@app/lib/swr/swr";
+import type {
+  GetBySpacesSummaryResponseBody,
+  GetSpaceConversationsResponseBody,
+  GetSpaceUnreadConversationsResponseBody,
+} from "@app/types/api/assistant/conversation/spaces";
 import type { LightConversationType } from "@app/types/assistant/conversation";
 import { useCallback, useMemo } from "react";
 import type { Fetcher } from "swr";
