+++ front/lib/api/assistant/conversation/forks.ts
@@ -55,12 +55,6 @@ export type CreateConversationForkErrorCode =
   | "invalid_request_error"
   | "unauthorized";
 
-export type PostConversationForkResponseBody = {
-  conversationId: string;
-  parentConversationTitle: string | null;
-  spaceId: string | null;
-};
-
 type CarriedAttachment = {
   carriedAttachment:
     | ContentFragmentInputWithFileIdType
