+++ front/types/api/assistant/conversation/forks.ts
@@ -1,0 +1,5 @@
+export type PostConversationForkResponseBody = {
+  conversationId: string;
+  parentConversationTitle: string | null;
+  spaceId: string | null;
+};
