+++ front/types/api/assistant/conversation/search.ts
