+++ front/types/api/assistant/conversation/spaces.ts
