+++ front/types/api/assistant/conversation/wakeups.ts
